adding scripts, fixing scripts, testing scripts under Ubuntu 16.04 x64; currently untested: ./single_installers/install_ompl.sh; several not-working-under-16.04 scripts are listed as such and have an exit 0 early-on to prevent a full run

+++ single_installers/install_ws4py.sh
@@ -18,38 +18,42 @@ RELATIVE_PATH="`dirname \"$0\"`"
 ABSOLUTE_PATH="`( cd \"$RELATIVE_PATH\" && pwd )`"
 echo "PATH of current script ($0) is: $ABSOLUTE_PATH"
 
-# find O/S codename (set to UCODENAME)
-source $ABSOLUTE_PATH/get_os_codename.sh
-
 #
 # parse input vars (set to appropriate vars or default vars)
 #
 source $ABSOLUTE_PATH/get_rv_su_wd_f.sh "$@"
 # when source'd, sets these vars at this level: ROSVERSION SCRIPTUSER WORKSPACEDIR FORCE
 
-
 #
 # check for installation
 #
 
-WS4PY_FOUND=`python -c "import pkg_resources; print(pkg_resources.get_distribution('ws4py').version)" | grep -m 1 -o "0.3.5" | wc -l`
-# pkg_resources should give '0.3.5'
+#WS4PY_VER="0.3.5"
+WS4PY_VER="0.4.3.dev0" # 2017-05-19
+WS4PY_FOUND=`python -c "import pkg_resources; print(pkg_resources.get_distribution('ws4py').version)" | grep -m 1 -o "$WS4PY_VER" | wc -l`
+# pkg_resources should give '$WS4PY_VER'
 # grep should find a match and repeat it
-# and wc -l should give 1 if ws4py 0.3.5 was found
+# and wc -l should give 1 if ws4py of $WS4PY_VER was found
 if [ $WS4PY_FOUND -eq 1 ]; then
-    echo "ws4py 0.3.5 already installed!"
+    echo "ws4py $WS4PY_VER already installed!"
 fi
 # longer test from within python:
 # >>> from ws4py.client.threadedclient import WebSocketClient
 # >>> testclient = WebSocketClient('ws://localhost:9090/')
+
+# exit script immediately if libraries are already installed
+if [ "$FORCE" != "-f" ] && [ $WS4PY_FOUND -eq 1 ]
+then
+    echo "Websocket-for-Python libraries already installed and up-to-date, exiting..."
+    exit 0
+fi
 
 #
 # run installation + upgrades
 #
 
 # update all packages, because "gah!" otherwise, especially for 'rosdep' stuff later
-sudo apt-get -y update
-sudo apt-get -y upgrade
+$ABSOLUTE_PATH/apt_upd_sys.sh
 
 sudo apt-get -y install wget curl # for wget and possible curl use below
 
@@ -67,7 +71,5 @@ cd "WebSocket-for-Python"
 if [ "$FORCE" == "-f" ] || [ $WS4PY_FOUND -eq 0 ]; then
     sudo python setup.py install
 fi
-
-
 
 echo "End of install_ws4py.sh script!"
